+++ Cargo.toml
@@ -14,4 +14,5 @@ path = "src/lib.rs"
 actix-web = "4.3.1"
 reqwest = "0.11.18"
 serde = { version = "1.0.163", features = ["derive"]}
+sqlx = { version = "0.6.3", default-features = false, features = ["runtime-tokio-rustls", "macros", "postgres", "uuid", "chrono", "migrate"] }
 tokio = { version = "1.28.1", features = ["macros", "rt-multi-thread"] }
