+++ src/lib.rs
@@ -1,29 +1,3 @@
-use std::net::TcpListener;
-
-use actix_web::{dev::Server, web, App, HttpResponse, HttpServer, Responder};
-
-async fn health_check() -> impl Responder {
-    HttpResponse::Ok()
-}
-
-#[derive(serde::Deserialize)]
-struct FormData {
-    email: String,
-    name: String,
-}
-
-async fn subscribe(form: web::Form<FormData>) -> impl Responder {
-    HttpResponse::Ok()
-}
-
-pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
-    let server = HttpServer::new(|| {
-        App::new()
-            .route("/health_check", web::get().to(health_check))
-            .route("/subscribe", web::post().to(subscribe))
-    })
-    .listen(listener)?
-    .run();
-
-    Ok(server)
-}
+pub mod configuration;
+pub mod routes;
+pub mod startup;
