+++ src/startup.rs
@@ -1,0 +1,19 @@
+use std::net::TcpListener;
+
+use crate::routes::{health_check, subscribe};
+use actix_web::dev::Server;
+use actix_web::web;
+use actix_web::App;
+use actix_web::HttpServer;
+
+pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
+    let server = HttpServer::new(|| {
+        App::new()
+            .route("/health_check", web::get().to(health_check))
+            .route("/subscribe", web::post().to(subscribe))
+    })
+    .listen(listener)?
+    .run();
+
+    Ok(server)
+}
